設定修正

+++ hooks/copy_native_files.js
@@ -20,5 +20,11 @@ module.exports = function (context) {
     "cpp"
   );
 
-  fs.copySync(srcDir, destDir);
+  if (!fs.existsSync(destDir)) {
+    fs.mkdirSync(destDir, { recursive: true });
+    console.log(`✨️Created directory: ${destDir}`);
+  } else {
+    fs.copySync(srcDir, destDir);
+    console.log(`✨️Copyed directory: ${destDir}`);
+  }
 };
